Desktop/poc/requirements.txt
agent-framework-ollama==1.0.0b251218
agent-framework-devui==1.0.0b251218